libps (_proc_stat_free): Fix memory leak
The thread_waits field was never freed; this change adds the code to do
so if the corresponding flag is set.

Users are also warned of the consequences, namely, that they should not
free process proc_stats before all their associated thread ones are
gone, since the latter may reference the former's thread_waits field.
However, with the normal interface (ie.  ps_context_free), all
proc_stats are freed at once, so this is not an issue.

* libps/procstat.c (_proc_stat_free): Free ps->thread_waits if
PSTAT_THREAD_WAITS flag is set.
* libps/ps.h (_proc_stat_free): Document that thread proc_stats must be
freed before process proc_stat can be.