libps (fetch_procinfo): Fix non-set flag PSTAT_TASK_EVENTS
The fetch_procinfo function failed to set the PSTAT_TASK_EVENTS when
those has been retreived. It should be noted that GNU Mach does not
support the TASK_EVENTS_INFO flavor, so this is of minor importance.

The new code also checks that all the requested proc server flags
related to a given proc_stat flags have been returned before setting the
proc_stat flag in question.

* libps/procstat.c (fetch_procinfo): Add map array containing `ps_flags_t' and
`procinfo' flags.  Use it in a loop to check for flags in `need' and `have' and
set them in `pi_flags'.  Likewise to check for flags in `pi_flags' and set them
in `have'.