Find the absolute path to <X11/keysymdef.h>
Specifying the header file relative to $(includeir) won't work because
it is defined in terms of prefix variable which is redefined during
installation to the destination directory.

And the dependency check will fail during 'make install' because the
header might not be in the destination tree.

* config.make.in (X11_KEYSYMDEF_H, SED): New variables.
* configure.in: Check for sed. Find X11/keysymdef.h absolute location
using the preprocessor and save it in X11_KEYSYMDEF_H variable.
* console-client/Makefile (kstoucs_map.c): Use X11_KEYSYMDEF_H to
depend on the header file.