Sync pager before clearing MAY_CACHE flag
Clearing MAY_CACHE flag on a pager initiates a memory object
termination if this one is not referenced anymore. If the object has a
significant number of dirty pages (i.e. a file recently created was
unlinked before diskfs periodical sync) this operation generates a lot
of stress on the translator. This is one of the most common sources
for thread storms.

Sync'ing the pager before clearing that flag ensures that there aren't
dirty pages in the object before its termination.

* ext2fs/pager.c (drop_pager_softrefs): Sync pager before clearing
MAY_CACHE flag.