tmpfs: add --size
Add the possibility to specify the size with the --size parameter;
this makes tmpfs more usable in fstab or Linuxish mount invocations,
since the size in such cases is a mount -o option, which gets translated
to a --foo translator argument.
The old way (specifying the size as the first argument) is left there;
although, if --size is passed then the first argument must be "tmpfs",
as it is what is passed by fstab/mount.

* tmpfs/tmpfs.c (OPT_SIZE): New macro.
(options): Add the "size" option.
(parse_opt): Use -1 to indicate when SIZE is not yet set.
<OPT_SIZE>: Handle case.
<ARGP_KEY_NO_ARGS>: Error out only when SIZE is not set.
<ARGP_KEY_ARGS>: Error out when SIZE is set and the argument is not "tmpfs".