Fix directory rename in firmlinked directory
checkpath browses directories up to look against renaming a directory into
itself. It used to assume being able to stop at the root of the filesystem.  But
some per-opens have a shadow_root, where recursion has to stop too.

This means checkpath does not check up to the root any more. This is fine, since
we already prevent cross-shadow-root renames very early in dir-rename.c, we
do not need to check outside of this shadow root.

* libdiskfs/dir-renamed.c (checkpath): Also stop at the per-open shadow_root.