hurd: add fsys_get_source
Return information about the source of the receiving filesystem.
If the concept of a source is applicable, SOURCE should refer to
the source of the receiving translator and should be a description
considered appropriate in the context of the translator.  For
example for the case of block device based filesystems, SOURCE
should be the file name of the underlying block device.

* hurd/fsys.defs: Add fsys_get_source.
* hurd/fsys_reply.defs: Add fsys_get_source.