init: Mark all of inits children and init itself as important
This is based on a fragment of Guillem Jovers patch presented here:

http://lists.gnu.org/archive/html/bug-hurd/2006-02/msg00081.html

It has been refreshed, updated and the copyright year is adjusted
properly. It has been complemented with the necessary features to
address the issues the original patch set out to address, namely
that killall5 freezes the proc translator before it tries to walk
over /proc/*/stat to decide which process to kill. Prior to this
patch (and the one marking the procfs server as important
process), killall5 would deadlock trying to walk over the proc
file system.

Ironically it would not have killed any process later on even if
it had the chance, since two values obtained from /proc/*/stat
are currently hardcoded to zero in our procfs. Patches addressing
the problem as a whole are prepared and will be sent as a follow
up.

* init/init.c (launch_core_servers): Mark init, auth, proc and fs
  servers as important.
(start_child): Mark the real init as important.
(S_startup_important_task): Mark exec server as important.