exec: remove the BFD code
This commit removes the parts of the exec server that were once using
the Binary File Descriptor library.

As I understand it, the BFD code stopped working because it uses stdio
streams and the glue code for that has never been ported to libio. The
code has been #ifdefed out ever since. It may or may not work in its
current state, so it is removed. If someone is interested, it can
always be recovered from the version control system.

* exec/exec.c: Remove all BFD related code.
* exec/priv.h: Likewise.
* TODO: Remove the corresponding item.