init: add a minimalist init program
This patch adds a minimalist init program.  It is somewhat lacking in
features, but is able to bring up a Hurd system with the runsystem and
rc scripts.  In fact, it roughly does what the former /hurd/init did,
modulo all the very early bootstrapping stuff and the startup
protocol.  It is started when all the essential servers are up and
running, so it can make use of most of the POSIX goodies, making its
implementation much simpler.

* Makefile (prog-subdirs): Add init.
* daemons/runsystem.sh: Generalize runsystem so that it can start any
init as specified on the kernel command line.  By default, it starts
/hurd/init.
* daemons/runsystem.hurd: This is a verbatim copy of runsystem.sh.  It
is started by /hurd/init.
* daemons/rc.sh: Do not start /hurd/mach-defpager as it is already
started in runsystem.sh.
* daemons/Makefile (SRCS): Add runsystem.hurd.
* init/Makefile: New file.
* init/init.c: Likewise.