startup: bind the startup server to /servers/startup
Previously, the Hurd (ab)used the fact that the startup server speaks
all protocols on its message port.  Any server that wished to register
for shutdown notifications would use proc_getmsgport to get a port to
the startup server.

This hardcodes the PID of /hurd/startup, and does not allow one to
point a server to ones own startup server (e.g. using remap).

Bind the startup server to /servers/startup instead.  Use this to
contact the startup server.

* exec/main.c (S_exec_init): Use /servers/startup.  Fall back to the
old method so that the system still boots when the node
/servers/startup is missing.
* hurd/paths.h (_SERVERS_STARTUP): New macro.
* libdiskfs/boot-start.c (diskfs_S_fsys_init): Use /servers/startup.
* libdiskfs/init-startup.c (_diskfs_init_completed): Likewise.
* pfinet/main.c (arrange_shutdown_notification): Likewise.
* startup/Makefile (OBJS): Add fsysServer.o.
* startup/startup.c (demuxer): Handle the fsys protocol.
(main): Bind to /servers/startup.
(S_fsys_getroot): Implement fsys_getroot.  Stub out the rest.