Revert "tmpfs: drop privileges in the tmpfs translator"
This reverts commit 7075bc11289355478659cd1610632a9299734afc.

Dropping the unix privileges of the tmpfs translator causes problems
with the passive translator startup. The function fshelp_fetch_root
creates an auth handle containing the processes original UID and
GID. But if the privileges were dropped before, this operation fails
with EPERM.

This commit also broke mkfifo(3), possibly for similar or related
reasons.

* tmpfs/tmpfs.c (main): Do not drop all privileges.