proc: fix a use-after-free error
If we have to create a new process group, we have to do this before
leaving the current one. The current process group is deallocated if
the process is the last process in that group. Likewise, if the
current group was the last group in the current groups session, the
session is deallocated.

Found using the Clang Static Analyzer.

* proc/pgrp.c (S_proc_setpgrp): Fix use-after-free error.