trans: improve the performance of dir_lookup in fakeroot
Previously any FS_RETRY_NORMAL requests were handed back to the
client.  Redo the lookup ourself instead.  This saves us the burden to
create a fake node for the intermediate step, hand it back to the
client and handle another request from the client.

With this change there is no need to fiddle with the permission bits
as it was previously done.

* trans/fakeroot.c (netfs_S_dir_lookup): Redo the lookup transparently
  for the user if FS_RETRY_NORMAL is requested.

fixup_improve_perf