trans/fakeroot: shutdown the translator if the last client is gone
Previously, fakeroot would not exit if a process outlived the original
process started by settrans.  This caused bugs like this:

% fakeroot-hurd /bin/sh -c 'sleep 1&' 2>&1 | tee
<hangs>

Fix this by exiting if the last client of fakeroot goes away.

If noone has a right to any control or protid port, noone can ever
reacquire any such right. So it is safe to shutdown the fakeroot
translator in that case.

* trans/fakeroot.c (fakeroot_netfs_release_protid): Shutdown the
translator if the last protid object is destroyed and no control port
is around either.