mach-defpager: fix warning about uninitialized variable
Previously, failure to look up the given partition was detected after
the loop by checking whether the loop ran over all existing
partitions.  Initialize part to NULL and check for that instead.  This
retains the behavior, but expresses it in a way the compiler
understands better.

* mach-defpager/default_pager.c (destroy_paging_partition): Initialize
part to NULL and check for it still being NULL after the loop.