mach-defpager: silently ignore requests to page to active partition
Currently, if mach-defpager is asked to page to an already active
partition, it ignores this request and returns success.  It does,
however, print a message about this to stdout.

This message might indicate to the user that there is some kind of a
problem with the configuration of the machine, even though the code in
new_partition clearly does not consider it an error to do such an
request.

* mach-defpager/default_pager.c (new_partition): Do not print a
message if requested to page to an already active partition.