ext2fs: fix type of inum
Previously, inum was of type int, whereas dino_ref expects ino_t.  On
Hurd/x86 the former is 32 bit wide, the latter 64.  If dino_ref is
inlined, this does not seem to pose a problem, but if ext2fs is
compiled with -O0, this most likely results in an invalid memory access.

* ext2fs/ialloc.c (ext2_alloc_inode): Use type ino_t for inum.