libdiskfs: remove the statistics code from the name cache
The current name cache lookup operation completes in O(n) time.  This
means that making the cache too large would decrease the performance.
Therefore it was required to tune the size, hence the need for
statistics.

We will use a data structure with worst case constant lookup times in
the future, removing the need to fine tune the cache size.

* libdiskfs/name-cache.c: Remove the statistics code from the name
cache.