libports: avoid realloc(3) corner case
If the size argument is 0, realloc may either return NULL, or return a
pointer that is only valid for use with free(3).  In either case, the
memory is freed.  So if realloc would return NULL (it does not on
GNU), the current code would double free p.

Found using the Clang Static Analyzer.

* libports/bucket-iterate.c (_ports_bucket_class_iterate): Avoid
calling realloc if no ports were matched.