hurd: fix semantic of file_get_children
When first introduced as fsys_get_children, it made sense to return
the list of children using paths relative to the root of the
filesystem that was queried.  Making the get_children method part of
the fsys protocol was a mistake that has since been corrected in
9366d6b2.

Instead of returning paths relative to the root of the translator,
return paths relative to the path of the receiving node.

This fixes a problem with the mtab translator.  Previously, the mtab
translator invoked on a target that was not the root directory of a
translator would compute invalid paths, e.g.:

/hurd/mtab: /any/path/servers/socket/26 No such file or directory

* hurd/fs.defs (file_get_children): Update comment.
* libfshelp/translator-list.c (fshelp_get_active_translators): Add
argument PREFIX.  Filter entries not beginning with PREFIX if
non-NULL, and omit PREFIX from the returned paths.
* libfshelp/fshelp.h (fshelp_get_active_translators): Update comment
accordingly.  Also clarify that both FILTER and PREFIX can be NULL.
* libdiskfs/file-get-children.c (diskfs_S_file_get_children): Update
comment, pass prefix to fshelp_get_active_translators.
* libnetfs/file-get-children.c (netfs_S_file_get_children): Likewise.