proc: call `startup_essential_task' earlier
Previously, the proc server did not call `startup_essential_task'
until it got the message port of the startup server using
`proc_setmsgport'.

Now that we have `/servers/startup', we can do this in main, before we
start our message service loop.

A complication arises because the traditional startup server is
single-threaded.  Handle this by tweaking startup not to bind itself
to `/servers/startup' before it is ready.

* proc/main.c (main): Try to lookup `/servers/startup' and send the
message here, or...
* proc/msg.c (S_proc_setmsgport): ... fall back to the old way here.
* proc/proc.h (startup_fallback): New variable.
* startup/startup.c (main): Move code installing ourself on `/servers/startup'
(install_as_translator): ... here.
(launch_core_servers): And use it here, just before we reply to `/hurd/auth'.