exec: fix setting the name of early servers
Previously, the exec server did not set the name of the servers
started before the proc server.  Instead this was done by the startup
server, but this was merely a workaround, missing notably the startup
server itself.

* exec/exec.c (set_name): New function.
(do_exec): Move the code setting the name to a new function, and also
call it if the proc server is not started yet.
* startup/startup.c (run): Drop hack.