trans/ifsock: convert to trivfs dynamic classes and buckets
libtrivfs contains two ways of managing more than one port class and
bucket.  There is the old way of using a statically allocated array
with explicit length, and the new way with dynamically allocated
vectors.

Converting all users to the new way of handling multiple classes
and/or buckets, we can simplify the code in libtrivfs.  In many cases,
the code will be simpler and more expressive for the user.

This also fixes a mild bug.  The classes and buckets given to
`trivfs_startup' end up in the dynamic vectors too, making the object
lookup code use the more complicated code path.

* trans/ifsock.c: Convert to dynamic classes and buckets.