trans/magic: convert to trivfs dynamic classes and buckets
libtrivfs contains two ways of managing more than one port class and
bucket.  There is the old way of using a statically allocated array
with explicit length, and the new way with dynamically allocated
vectors.

Converting all users to the new way of handling multiple classes
and/or buckets, we can simplify the code in libtrivfs.  In many cases,
the code will be simpler and more expressive for the user.

This also fixes a severe bug.  As no classes are given to
`trivfs_startup', they are created and inserted into the dynamic
vector of classes.  The server function `trivfs_S_fsys_forward',
however, used the first item of the previously used static array,
which is NULL.  This circumvented the typecheck.

* trans/magic.c: Convert to dynamic classes and buckets.