libihash: rehash if effective load exceeds the threshold
Previously, if a hash table was not growing anymore but entries were
regularly replaced, the tombstones would accumulate slowing down
lookups and insertions.  A possible solution is to rehash the table if
the effective load exceeds the configured threshold.  The effective
load also takes tombstones into account.

* libihash/ihash.c (hurd_ihash_locp_add): Use the effective load.
(hurd_ihash_add): Likewise.  Use the load to decide whether we want to
enlarge the table, otherwise we merely rehash.