Fix loading small pic programs
It happens that the link script for ld.so contains a hole, which might thus
leave an empty page between the text and the data.  When loading a small pic
program, its text would then fit in there, and loading the data right after
it would fail.  We here rather force all pic loads to be mapped
contiguously, starting from the place that was allocated for the first pic
load.

* exec/exec.c (load_section): Return the address of the end of the section.
(load): Take the address to be used for loading pic objects as parameter,
force pic objects there if it is not zero, and compute and return the
address to be used for the next pic object.
(do_exec): Pass addresses for pic loads between calls to load().