proc: Remove erroneous process entry.
Previously, there was a spurious copy of PID 1 in the process table,
usually PID 8.  PID 1, usually a sysvinit compatible process, needs to
be PID 1 even though it is not actually the first process on the Hurd.
To this end, PID 1 is reserved, and the task is later supplied to proc
via proc_set_init_task once startup starts sysvinit.

Fixes 8d16db0cc28b2d911aee918d5c3582ad29ddfeed.

* proc/mgt.c (S_proc_set_init_task): Check if sysvinit's task already
made it into the process table, and remove it if so.