libfshelp/get-identity.c: Avoid spurious ihash removal
Reported by Brent W. Baccala.

While some thread has converted a hardref to a weakref and tries to release
the hash weakref, another thread might reacquire a hardref, and then convert
it to a weakref and try to release it. We thus have to make sure that we
really have the last weakref before removing from the hash.

* libfshelp/get-identity.c (id_clean): Also check that there are only
two weak refs left.