Revert "libdiskfs: Shutdown pagers on startup_dosync."
This reverts commit 89a49ec8b3a847f3b770cff6a4a80b98b94ad7bf.

There was a typo in diskfs_shutdown (FSYS_GOAWAY_FORCE || FSYS_GOAWAY_RECURSE)
which was making it FSYS_GOAWAY_NOWAIT, and thus not actually syncing if
there happens to be clients (it was then returning EBUSY). Fixing this into
FSYS_GOAWAY_FORCE | FSYS_GOAWAY_RECURSE then makes it wait for clients (and
NOWAIT is not supported), and there are very often some left, thus not
syncing either.