pflocal: Fix referencing connection queue entries.
As asserted in connq_destroy, for each entry in the queue we are supposed to
keep a reference to the socket that contains the queue.  So we need to
keep it when connecting and release it when accepting.

* pflocal/socket.c (S_socket_connect): Do not deref the peer socket when
sock_connect succeeded.
(S_socket_accept): Deref the socket when the accept succeeded.